:lipstick: rename files to camel case
