+++ personalization-webcomponents/src/checklist-detail.ce.vue
@@ -66,8 +66,8 @@ import { computed, onMounted, ref } from "vue";
 import DummyChecklistService from "@/api/dummyservice/DummyChecklistService.ts";
 import ChecklistHeader from "@/components/ChecklistHeader.vue";
 import ChecklistList from "@/components/ChecklistList.vue";
-import SkeletonLoader from "@/components/common/skeleton-loader.vue";
-import { QUERY_PARAM_CHECKLIST_ID } from "@/util/constants.ts";
+import SkeletonLoader from "@/components/common/SkeletonLoader.vue";
+import { QUERY_PARAM_CHECKLIST_ID } from "@/util/Constants.ts";
 
 const checklist = ref<DummyChecklist>();
 const loading = ref(true);
