+++ personalization-webcomponents/src/components/ChecklistCheckedCircle.vue
