+++ personalization-webcomponents/src/components/common/SkeletonLoader.vue
