+++ personalization-webcomponents/src/util/Constants.ts
